feat: #44. conent 수강 등록 요청 api 설계 완료

+++ backend/single/src/main/java/com/scf/single/domain/entity/ContentCheckUser.java
@@ -27,7 +27,7 @@ public class ContentCheckUser {
     private int complete;
 
     @Column(name = "member_id")
-    private int memberId;
+    private Long memberId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "content_id", nullable = false)
